chore(clerk-js,nextjs,shared): Move devBrowser to shared
