+++ packages/clerk-js/src/core/constants.ts
@@ -1,9 +1,5 @@
 export const PRESERVED_QUERYSTRING_PARAMS = ['after_sign_in_url', 'after_sign_up_url', 'redirect_url'];
 
-export const DEV_BROWSER_SSO_JWT_KEY = 'clerk-db-jwt';
-export const DEV_BROWSER_SSO_JWT_PARAMETER = '__dev_session';
-export const DEV_BROWSER_SSO_JWT_HTTP_HEADER = 'Clerk-Cookie';
-
 export const CLERK_MODAL_STATE = '__clerk_modal_state';
 export const CLERK_SYNCED = '__clerk_synced';
 export const CLERK_SATELLITE_URL = '__clerk_satellite_url';
