+++ packages/clerk-js/src/utils/cookies/devBrowser.ts
@@ -1,5 +1,4 @@
 import { createCookieHandler } from '@clerk/shared/cookie';
-
-import { DEV_BROWSER_JWT_MARKER } from '../devBrowser';
+import { DEV_BROWSER_JWT_MARKER } from '@clerk/shared/devBrowser';
 
 export const devBrowserCookie = createCookieHandler(DEV_BROWSER_JWT_MARKER);
