+++ packages/clerk-js/src/utils/index.ts
@@ -2,7 +2,6 @@ export * from './beforeUnloadTracker';
 export * from './componentGuards';
 export * from './cookies';
 export * from './dynamicParamParser';
-export * from './devBrowser';
 export * from './email';
 export * from './encoders';
 export * from './errors';
