+++ packages/shared/src/index.ts
@@ -30,3 +30,10 @@ export * from './proxy';
 export * from './underscore';
 export * from './url';
 export { createWorkerTimers } from './workerTimers';
+export {
+  DEV_BROWSER_SSO_JWT_PARAMETER,
+  DEV_BROWSER_JWT_MARKER,
+  getDevBrowserJWTFromURL,
+  setDevBrowserJWTInURL,
+  getDevBrowserJWTFromResponse,
+} from './devBrowser';
